jhrg/HYRAX-1828-Redis-file-cache (#1)
Working unit tests.

Added a blocking read test. The blocking writer
test did not actually test that the code would block
but instead tested only that if you ask it to wait
long enough, it would work. Now it tests both
things. This might not work with a remote Redis
server where it takes longer to communicate. IDK

Fixed the LRU eviction test.