jhrg/hyrax 1829 add cicd support (#2)
Added a simple travis.yml file
Added CppUnit and the 'never' travis branch.
Added debug output to CMakeLists in Cpp/unit-tests
Tests added to CMakeLists and wired up so they work with ctest
Added the call to ctest to the travis build.

Build the build and the unit tests are now run. The tests use the 
Travis built in Redis server.